separated cylinder geometries to separate file, added radius calculation to vessel geometry, added vessel surface elasticity to maintain vessel shape
